+++ studyapp/app/(tabs)/_layout.tsx
@@ -1,4 +1,4 @@
-import { Tabs, Stack } from 'expo-router';
+import { Tabs } from 'expo-router';
 import React from 'react';
 import { TabBarIcon } from '@/components/navigation/TabBarIcon';
 import { Colors } from '@/constants/Colors';
