+++ studyapp/package.json
@@ -19,6 +19,7 @@
     "@expo/vector-icons": "^14.0.2",
     "@gluestack-ui/avatar": "^0.1.17",
     "@gluestack-ui/button": "^1.0.7",
+    "@gluestack-ui/input": "^0.1.31",
     "@gluestack-ui/nativewind-utils": "^1.0.23",
     "@gluestack-ui/overlay": "^0.1.15",
     "@gluestack-ui/toast": "^1.0.7",
